producer/requirements.txt
kafka-python==2.2.15
Faker==37.6.0